+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/extractor/TokenBrokerResolver.java
@@ -5,10 +5,7 @@ import javax.servlet.http.HttpServletRequest;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
-import java.util.Enumeration;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -33,8 +30,8 @@ import com.sap.cloud.security.xsuaa.tokenflows.XsuaaTokenFlows;
  * Analyse authentication header and obtain token from UAA
  *
  * For using this feature also in multi tenancy mode request-parameter
- * {@code X-Identity-Zone-Subdomain} must be set or the
- * AuthenticationInformationExtractor needs to be implemented).
+ * {@code X-Identity-Zone-Subdomain} must be set
+ * (or the AuthenticationInformationExtractor needs to be implemented).
  *
  */
 public class TokenBrokerResolver implements BearerTokenResolver {
@@ -119,7 +116,7 @@ public class TokenBrokerResolver implements BearerTokenResolver {
 		try {
 			return extractToken(request);
 		} catch (TokenBrokerException e) {
-			logger.warn("Error obtaining token:" + e.getMessage(), e);
+			logger.warn("Error obtaining token: " + e.getMessage(), e);
 			return null;
 		}
 	}
@@ -131,11 +128,14 @@ public class TokenBrokerResolver implements BearerTokenResolver {
 
 		String oauthTokenUrl = getOAuthTokenUrl(request);
 
-		Enumeration<String> authHeaderValues = request.getHeaders(AUTH_HEADER);
+		ClientCredentials clientCredentials = new ClientCredentials(configuration.getClientId(),
+				configuration.getClientSecret());
 		for (AuthenticationMethod credentialType : authenticationMethods) {
-			String token = getBrokerToken(credentialType, authHeaderValues, oauthTokenUrl, request);
-			if (!StringUtils.isEmpty(token)) {
-				return token;
+			for (String authHeaderValue : Collections.list(request.getHeaders(AUTH_HEADER))) {
+				String token = getBrokerToken(credentialType, authHeaderValue, oauthTokenUrl, clientCredentials);
+				if (!StringUtils.isEmpty(token)) {
+					return token;
+				}
 			}
 		}
 		return null;
@@ -165,13 +165,10 @@ public class TokenBrokerResolver implements BearerTokenResolver {
 		return oauthTokenUrl;
 	}
 
-	private String getBrokerToken(AuthenticationMethod credentialType, Enumeration<String> authHeaderValues,
-			String oauthTokenUrl, final HttpServletRequest request) throws TokenBrokerException {
-		ClientCredentials clientCredentials = new ClientCredentials(configuration.getClientId(),
-				configuration.getClientSecret());
-		while (authHeaderValues.hasMoreElements()) {
-			String authHeaderValue = authHeaderValues.nextElement();
-			switch (credentialType) {
+	private String getBrokerToken(AuthenticationMethod credentialType, String authHeaderValue,
+			String oauthTokenUrl, ClientCredentials clientCredentials) throws TokenBrokerException {
+				logger.info("AuthHeaderValue: {}", authHeaderValue);
+		switch (credentialType) {
 			case OAUTH2:
 				return extractAuthenticationFromHeader(AUTH_BEARER, authHeaderValue);
 			case BASIC:
@@ -217,7 +214,6 @@ public class TokenBrokerResolver implements BearerTokenResolver {
 				break;
 			default:
 				return null;
-			}
 		}
 		return null;
 	}