+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/extractor/UaaTokenBroker.java
@@ -41,7 +41,7 @@ public class UaaTokenBroker implements TokenBroker {
 					URI.create(tokenURL), new ClientCredentials(clientId, clientSecret), null, null).getAccessToken();
 		} catch (OAuth2ServiceException ex) {
 			logger.warn("Cannot obtain Client Credentials Access Token for clientId {}.", clientId);
-			throw new TokenBrokerException("Cannot obtain Client Credentials Access Token from given clientId.", ex);
+			throw new TokenBrokerException("Cannot obtain Client Credentials Access Token for given clientId.", ex);
 		}
 	}
 
@@ -53,8 +53,8 @@ public class UaaTokenBroker implements TokenBroker {
 					URI.create(tokenURL), new ClientCredentials(clientId, clientSecret), username, password, null, null)
 					.getAccessToken();
 		} catch (OAuth2ServiceException ex) {
-			logger.warn("Cannot obtain Token from given user / password.");
-			throw new TokenBrokerException("Cannot obtain Token from given user / password.", ex);
+			logger.warn("Cannot obtain Token for user / password {}.", username);
+			throw new TokenBrokerException("Cannot obtain Token for given user / password.", ex);
 		}
 	}