+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 - [token-client] [OidcConfigurationServiceWithCache](/token-client/src/main/java/com/sap/cloud/security/xsuaa/client/OidcConfigurationServiceWithCache.java) supports basically Open-id Configuration endpoints as documented [here](https://docs.cloudfoundry.org/api/uaa/version/74.4.0/index.html#openid-connect-flow).
 - [token-client] [OAuth2TokenKeyServiceWithCache](/token-client/src/main/java/com/sap/cloud/security/xsuaa/client/OAuth2TokenKeyServiceWithCache.java) supports JWKS endpoint with cache of identity service as documented [here](https://docs.cloudfoundry.org/api/uaa/version/74.4.0/index.html#token-keys).
 
+## 2.3.1
+- [spring-xsuaa] Fix vulnerability issues and increased Spring versions.
+- [spring-xsuaa] **Bug fix** in TokenBrokerResolver: Second configured authentication method was ignored.
+
 ## 2.3.0
 - Issue: Spring tests fail with version `2.2.0`, when auto-configuration is disabled and no `RestOperations` bean is specified.
 - [token-client] Supports basically JWT Bearer Token Grant as documented [here](https://docs.cloudfoundry.org/api/uaa/version/74.4.0/index.html#jwt-bearer-token-grant).